fix(index.d.ts): fix getPixelWorldHeightAtCoord missing parameter
